basic gui
had some free time so i played around with swing to figure out how it works, this machine is badly made because i wasn't trying to make it well for now - it shows how you can respond to events like pressing buttons, i added general features like a text box for the display and an event that changes the display/title, you can probably implement contact with the vending machine software fairly easily. generally looking at this will give you a good idea of what the GUI is going to look like later on. Only 2 of the buttons and the loonie cause events for now. I hope the vending machine looks familiar to some of you.